refactor(gui-v2): rename durationHelper -> durationUtils
