+++ packages/nc-gui-v2/utils/durationUtils.ts
