+++ packages/nc-gui-v2/utils/index.ts
@@ -1,7 +1,7 @@
 export * from './colorsUtils'
 export * from './dateTimeUtils'
 export * from './deepCompare'
-export * from './durationHelper'
+export * from './durationUtils'
 export * from './errorUtils'
 export * from './fileUtils'
 export * from './filterUtils'
